+++ src/HCI/typingtest.cpp
@@ -12,10 +12,12 @@ TypingTest::TypingTest(QWidget * parent) : QWidget(parent) {
    text->setWordWrap(true);
    write = new Typing(text->text(), 0);
 
-   connect(write,SIGNAL(finished(const QVector<int>&, const QVector<int>&,
-				 const QVector<int>&)),
-	   this, SLOT(calculate(const QVector<int>&, const QVector<int>&,
-				const QVector<int>&))); 
+   connect(write,SIGNAL(finished(const QVector<long long>&,
+				 const QVector<long long>&,
+				 const QVector<long long>&)),
+	   this, SLOT(calculate(const QVector<long long>&,
+				const QVector<long long>&,
+				const QVector<long long>&))); 
 
    QVBoxLayout *layout = new QVBoxLayout(this);
    layout->addWidget(explain);
@@ -26,8 +28,9 @@ TypingTest::TypingTest(QWidget * parent) : QWidget(parent) {
    
 }
 
-void TypingTest::calculate(const QVector<int> &key, const QVector<int> &press,
-	       const QVector<int> &release) {
+void TypingTest::calculate(const QVector<long long> &key,
+			   const QVector<long long> &press,
+			   const QVector<long long> &release) {
    int latest = 0;
    analysis.addKeyFeature(key, press, release);
    if(analysis.isTired(latest)) {
