+++ src/HCI/typingtest.h
@@ -24,8 +24,8 @@ class TypingTest : public QWidget {
    TiredAnalysis analysis;
 
    public slots:
-   void calculate(const QVector<int>&, const QVector<int>&,
-		  const QVector<int>&);
+   void calculate(const QVector<long long>&, const QVector<long long>&,
+		  const QVector<long long>&);
 };
 
 #endif
